clara_7.0/requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.6
pydantic==2.9.1
python-multipart==0.0.9
jinja2==3.1.4
pypdf==4.3.1
pdfminer.six==20231228
pandas==2.2.2
pyyaml==6.0.2
rich==13.8.1
structlog==24.4.0
aiofiles==24.1.0
starlette==0.38.5
httpx==0.27.0
# Optional OCR; handled gracefully if missing
pytesseract==0.3.13
Pillow==10.4.0
